update: requirement detail view
